requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
Django==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
djongo==1.3.6
dnspython==2.7.0
exceptiongroup==1.3.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
mongoengine==0.29.1
multidict==6.6.3
openai==0.28.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==3.12.3
python-decouple==3.8
pytz==2025.2
requests==2.32.4
sniffio==1.3.1
sqlparse==0.2.4
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
